Fix build with libcurl
Add a patch that fixes 'autoreconf' so that the m4 directory is properly
read for autoconf to produce a valid configure script.

Also, query the libcurl cflags and libs with pkg-config instead of
curl-config. This avoids an issue with curl-config --libs listing
libraries private to libcurl (e.g. libnghttp2), since R does not need to
link to those directly and therefore does not need them for the build.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>